% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{atx_hex_to_raw}
\alias{atx_hex_to_raw}
\title{Hex string -> raw vector.}
\usage{
atx_hex_to_raw(h)
}
\arguments{
\item{h}{A hex string (even length).}
}
\value{
A raw vector.
}
\description{
Hex string -> raw vector.
}
\keyword{internal}
