% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{atx_raw_to_hex}
\alias{atx_raw_to_hex}
\title{Raw vector -> lowercase hex string.}
\usage{
atx_raw_to_hex(r)
}
\arguments{
\item{r}{A raw vector.}
}
\value{
A hex string.
}
\description{
Raw vector -> lowercase hex string.
}
\keyword{internal}
