% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didkey.R
\name{atx_verification_method}
\alias{atx_verification_method}
\title{Full verification method for a \code{did:key} (\code{<did>#<multibase>}).}
\usage{
atx_verification_method(did)
}
\arguments{
\item{did}{A \code{did:key:z...} string.}
}
\value{
The verification-method string.
}
\description{
Full verification method for a \code{did:key} (\code{<did>#<multibase>}).
}
