% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didkey.R
\name{.ED25519_MULTICODEC_PREFIX}
\alias{.ED25519_MULTICODEC_PREFIX}
\title{Ed25519 multicodec prefix (\code{0xed 0x01}) for \code{ed25519-pub}.}
\usage{
.ED25519_MULTICODEC_PREFIX
}
\description{
Ed25519 multicodec prefix (\code{0xed 0x01}) for \code{ed25519-pub}.
}
\keyword{internal}
