% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base58.R
\name{atx_base58_decode}
\alias{atx_base58_decode}
\title{Decode a base58btc string to raw bytes.}
\usage{
atx_base58_decode(s)
}
\arguments{
\item{s}{A length-1 character string.}
}
\value{
A raw vector.
}
\description{
Decode a base58btc string to raw bytes.
}
\keyword{internal}
